Moved number of rows and columns to profile and persisted to file Fixed Ping Pong messages between pi and app Fixed issue with profile json file being blank when first created Fixed double input spinner being really long.... like really long... Removed left panel. wasn't needed

+++ src/main/java/dev/theturkey/pideckapp/ui/ActionPanel.java
@@ -129,7 +129,7 @@ public class ActionPanel extends JPanel
 					});
 					break;
 				case DOUBLE:
-					component = new JSpinner(new SpinnerNumberModel(Float.parseFloat(val), Integer.MIN_VALUE, Float.MAX_VALUE, 0.01));
+					component = new JSpinner(new SpinnerNumberModel(Float.parseFloat(val), Integer.MIN_VALUE, Integer.MAX_VALUE, 0.01));
 					((JSpinner) component).addChangeListener(e ->
 					{
 						action.setPropsValue(prop.key, String.valueOf(((JSpinner) component).getModel().getValue()));
@@ -138,7 +138,6 @@ public class ActionPanel extends JPanel
 					break;
 				case BOOLEAN:
 					component = new JCheckBox("", Boolean.parseBoolean(val));
-
 					((JCheckBox) component).addItemListener(e ->
 					{
 						action.setPropsValue(prop.key, String.valueOf(((JCheckBox) component).isSelected()));
